requirements.txt
flask
flask_pymongo
flask_bcrypt
flask_login
python-dotenv
openai